backend/requirements.txt
# ── API server ───────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9

# ── LangChain core ───────────────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# ── LLM providers (install only what you use) ────────────────────────────────
langchain-ollama>=0.2.0
langchain-openai>=0.2.0          # also used for vLLM OpenAI-compat endpoint
langchain-huggingface>=0.1.0
transformers>=4.44.0
torch>=2.3.0
accelerate>=0.33.0

# ── Embeddings ───────────────────────────────────────────────────────────────
sentence-transformers>=3.1.0
fastembed>=0.3.0

# ── Vector stores ────────────────────────────────────────────────────────────
langchain-chroma>=0.1.4
chromadb>=0.5.0
langchain-qdrant>=0.2.0
qdrant-client>=1.11.0
langchain-milvus>=0.1.5
pymilvus>=2.4.0
langchain-mongodb>=0.3.0
pymongo>=4.9.0
langchain-neo4j>=0.3.0
neo4j>=5.24.0

# ── Parser ───────────────────────────────────────────────────────────────────
pymupdf>=1.24.0
pymupdf4llm>=0.0.17
pillow>=10.4.0
opencv-python-headless>=4.10.0

# ── Misc ─────────────────────────────────────────────────────────────────────
numpy>=1.26.0

# ── OCR (optional — uncomment if using scanned PDFs) ─────────────────────────
# doclayout-yolo
# paddleocr

# ── ONNX LLM (optional) ──────────────────────────────────────────────────────
# optimum[onnxruntime]
